pyproject.toml
[project]
name = "chico"
version = "0.1.0"
description = "Agent-native configuration control plane"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "PyGithub>=2.0",
]

[project.scripts]
chico = "chico.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "ty>=0.0.1a1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=chico --cov-report=term-missing --cov-fail-under=100"

[tool.coverage.run]
branch = true
source = ["chico"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "E501", # line too long — handled by formatter
    "TCH",  # type-checking blocks add boilerplate without benefit at this scale
]

[tool.ruff.lint.isort]
known-first-party = ["chico"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
